requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-RESTful==0.3.8
getpass3==1.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
pycodestyle==2.6.0
pylint==2.6.0
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
